refactor(services): migrate utility functions and services to shared/
Phase 1-2b (#2393): Migrate utility functions:
- actor_support.py → shared/actors.py
- artifact_parser.py → shared/artifacts.py
- timeline_parser.py → shared/timeline.py
- role_policy_helpers.py → shared/roles.py

Phase 1-2c (#2394): Migrate service classes:
- signature_service.py → shared/signatures.py
- loc_service.py → shared/loc.py
- version_service.py → shared/versions.py
- label_service.py → shared/label_service.py

All backward compatibility maintained via re-export shims.
Updated shared/__init__.py with lazy imports.

Closes #2393
Closes #2394

Related to #2099 (services subpackage migration epic)
