+++ src/vibe3/services/shared/__init__.py
@@ -3,6 +3,12 @@
 from typing import TYPE_CHECKING, Any
 
 if TYPE_CHECKING:
+    from vibe3.services.shared.actors import (
+        extract_role_from_actor,
+        format_agent_actor,
+        resolve_actor_backend_model,
+    )
+    from vibe3.services.shared.artifacts import ArtifactParser
     from vibe3.services.shared.branches import (
         resolve_branch_and_issue,
         resolve_branch_arg,
@@ -16,6 +22,7 @@ if TYPE_CHECKING:
         LabelAnomaly,
         collect_label_anomalies,
     )
+    from vibe3.services.shared.label_service import LabelService
     from vibe3.services.shared.labels import (
         ORCHESTRA_GOVERNED_LABEL,
         classify_dispatch_eligibility,
@@ -32,6 +39,7 @@ if TYPE_CHECKING:
         normalize_labels,
         should_skip_from_queue,
     )
+    from vibe3.services.shared.loc import LocService, LOCStats
     from vibe3.services.shared.paths import (
         GitClientProtocol,
         check_ref_exists,
@@ -43,8 +51,22 @@ if TYPE_CHECKING:
         resolve_ref_path,
         sanitize_event_detail_paths,
     )
+    from vibe3.services.shared.roles import get_role_block_function
+    from vibe3.services.shared.signatures import (
+        AI_ASSISTANT_ACTORS,
+        WORKFLOW_ACTOR,
+        SignatureService,
+    )
+    from vibe3.services.shared.timeline import parse_timeline_from_comments
+    from vibe3.services.shared.versions import VersionService
 
 __all__ = [
+    # actors
+    "resolve_actor_backend_model",
+    "format_agent_actor",
+    "extract_role_from_actor",
+    # artifacts
+    "ArtifactParser",
     # paths
     "GitClientProtocol",
     "check_ref_exists",
@@ -61,6 +83,7 @@ __all__ = [
     "record_error",
     # labels
     "LabelAnomaly",
+    "LabelService",
     "classify_dispatch_eligibility",
     "clean_old_state_labels",
     "collect_label_anomalies",
@@ -76,12 +99,32 @@ __all__ = [
     "normalize_labels",
     "ORCHESTRA_GOVERNED_LABEL",
     "should_skip_from_queue",
+    # loc
+    "LOCStats",
+    "LocService",
     # branches
     "resolve_branch_and_issue",
     "resolve_branch_arg",
+    # roles
+    "get_role_block_function",
+    # signatures
+    "AI_ASSISTANT_ACTORS",
+    "WORKFLOW_ACTOR",
+    "SignatureService",
+    # timeline
+    "parse_timeline_from_comments",
+    # versions
+    "VersionService",
 ]
 
 _SYMBOL_MODULES = {
+    # actors
+    "resolve_actor_backend_model": "vibe3.services.shared.actors",
+    "format_agent_actor": "vibe3.services.shared.actors",
+    "extract_role_from_actor": "vibe3.services.shared.actors",
+    # artifacts
+    "ArtifactParser": "vibe3.services.shared.artifacts",
+    # paths
     "GitClientProtocol": "vibe3.services.shared.paths",
     "check_ref_exists": "vibe3.services.shared.paths",
     "find_worktree_path_for_branch": "vibe3.services.shared.paths",
@@ -91,10 +134,13 @@ _SYMBOL_MODULES = {
     "ref_to_handoff_cmd": "vibe3.services.shared.paths",
     "resolve_ref_path": "vibe3.services.shared.paths",
     "sanitize_event_detail_paths": "vibe3.services.shared.paths",
+    # errors
     "has_recent_specific_error": "vibe3.services.shared.errors",
     "record_dispatch_failure_if_unexpected": "vibe3.services.shared.errors",
     "record_error": "vibe3.services.shared.errors",
+    # labels
     "LabelAnomaly": "vibe3.services.shared.label_anomalies",
+    "LabelService": "vibe3.services.shared.label_service",
     "classify_dispatch_eligibility": "vibe3.services.shared.labels",
     "clean_old_state_labels": "vibe3.services.shared.labels",
     "collect_label_anomalies": "vibe3.services.shared.label_anomalies",
@@ -110,8 +156,22 @@ _SYMBOL_MODULES = {
     "normalize_labels": "vibe3.services.shared.labels",
     "ORCHESTRA_GOVERNED_LABEL": "vibe3.services.shared.labels",
     "should_skip_from_queue": "vibe3.services.shared.labels",
+    # loc
+    "LOCStats": "vibe3.services.shared.loc",
+    "LocService": "vibe3.services.shared.loc",
+    # branches
     "resolve_branch_and_issue": "vibe3.services.shared.branches",
     "resolve_branch_arg": "vibe3.services.shared.branches",
+    # roles
+    "get_role_block_function": "vibe3.services.shared.roles",
+    # signatures
+    "AI_ASSISTANT_ACTORS": "vibe3.services.shared.signatures",
+    "WORKFLOW_ACTOR": "vibe3.services.shared.signatures",
+    "SignatureService": "vibe3.services.shared.signatures",
+    # timeline
+    "parse_timeline_from_comments": "vibe3.services.shared.timeline",
+    # versions
+    "VersionService": "vibe3.services.shared.versions",
 }
